fix: duplicate @tiptap/core import in footnote.ts, split formatting-toolbar.ts under 200 lines
Merged the separate `import type { NodeViewRendererProps }` into the existing
`@tiptap/core` value import in footnote.ts. Extracted ToolbarButton interface,
buildToolbarButtons, KB_HINTS, and buildButtonTitle into formatting-toolbar-actions.ts,
reducing formatting-toolbar.ts from 236 lines to 109 lines.
